+++ spec/sleeping_king_studios/tools/assertions_spec.rb
@@ -115,6 +115,125 @@ RSpec.describe SleepingKingStudios::Tools::Assertions do
     end
   end
 
+  describe '#assert_blank' do
+    let(:error_message) { 'value must be nil or empty' }
+
+    def assert
+      assertions.assert_blank(value, **options)
+    end
+
+    it 'should define the method' do
+      expect(assertions)
+        .to respond_to(:assert_blank)
+        .with(1).argument
+        .and_keywords(:as, :error_class, :message)
+    end
+
+    describe 'with nil' do
+      let(:value) { nil }
+
+      include_examples 'should not raise an exception'
+    end
+
+    describe 'with an Object' do
+      let(:value) { Object.new.freeze }
+
+      include_examples 'should raise an exception'
+    end
+
+    describe 'with an empty value' do
+      let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+      include_examples 'should not raise an exception'
+    end
+
+    describe 'with a non-empty value' do
+      let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+      include_examples 'should raise an exception'
+    end
+
+    wrap_context 'with as: value' do
+      let(:error_message) { "#{as} must be nil or empty" }
+
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+        include_examples 'should raise an exception'
+      end
+    end
+
+    wrap_context 'with error_class: value' do # rubocop:disable RSpec/RepeatedExampleGroupBody
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+        include_examples 'should raise an exception'
+      end
+    end
+
+    wrap_context 'with message: value' do # rubocop:disable RSpec/RepeatedExampleGroupBody
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+        include_examples 'should raise an exception'
+      end
+    end
+  end
+
   describe '#assert_boolean' do
     let(:error_message) { 'value must be true or false' }
 
@@ -1357,6 +1476,125 @@ RSpec.describe SleepingKingStudios::Tools::Assertions do
     end
   end
 
+  describe '#assert_presence' do
+    let(:error_message) { "value can't be blank" }
+
+    def assert
+      assertions.assert_presence(value, **options)
+    end
+
+    it 'should define the method' do
+      expect(assertions)
+        .to respond_to(:assert_presence)
+        .with(1).argument
+        .and_keywords(:as, :error_class, :message)
+    end
+
+    describe 'with nil' do
+      let(:value) { nil }
+
+      include_examples 'should raise an exception'
+    end
+
+    describe 'with an Object' do
+      let(:value) { Object.new.freeze }
+
+      include_examples 'should not raise an exception'
+    end
+
+    describe 'with an empty value' do
+      let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+      include_examples 'should raise an exception'
+    end
+
+    describe 'with a non-empty value' do
+      let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+      include_examples 'should not raise an exception'
+    end
+
+    wrap_context 'with as: value' do
+      let(:error_message) { "#{as} can't be blank" }
+
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+        include_examples 'should not raise an exception'
+      end
+    end
+
+    wrap_context 'with error_class: value' do # rubocop:disable RSpec/RepeatedExampleGroupBody
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+        include_examples 'should not raise an exception'
+      end
+    end
+
+    wrap_context 'with message: value' do # rubocop:disable RSpec/RepeatedExampleGroupBody
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+        include_examples 'should not raise an exception'
+      end
+    end
+  end
+
   describe '#validate' do
     let(:error_class)   { ArgumentError }
     let(:error_message) { 'block returned a falsy value' }
@@ -1396,6 +1634,100 @@ RSpec.describe SleepingKingStudios::Tools::Assertions do
         let(:block) { -> { :ok } }
 
         include_examples 'should not raise an exception'
+      end
+    end
+  end
+
+  describe '#validate_blank' do
+    let(:error_class)   { ArgumentError }
+    let(:error_message) { 'value must be nil or empty' }
+
+    def assert
+      assertions.validate_blank(value, **options)
+    end
+
+    it 'should define the method' do
+      expect(assertions)
+        .to respond_to(:validate_blank)
+        .with(1).argument
+        .and_keywords(:as, :message)
+    end
+
+    describe 'with nil' do
+      let(:value) { nil }
+
+      include_examples 'should not raise an exception'
+    end
+
+    describe 'with an Object' do
+      let(:value) { Object.new.freeze }
+
+      include_examples 'should raise an exception'
+    end
+
+    describe 'with an empty value' do
+      let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+      include_examples 'should not raise an exception'
+    end
+
+    describe 'with a non-empty value' do
+      let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+      include_examples 'should raise an exception'
+    end
+
+    wrap_context 'with as: value' do
+      let(:error_message) { "#{as} must be nil or empty" }
+
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+        include_examples 'should raise an exception'
+      end
+    end
+
+    wrap_context 'with message: value' do
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+        include_examples 'should raise an exception'
       end
     end
   end
@@ -2460,6 +2792,100 @@ RSpec.describe SleepingKingStudios::Tools::Assertions do
 
       describe 'with a Symbol' do
         let(:value) { :symbol }
+
+        include_examples 'should not raise an exception'
+      end
+    end
+  end
+
+  describe '#validate_presence' do
+    let(:error_class)   { ArgumentError }
+    let(:error_message) { "value can't be blank" }
+
+    def assert
+      assertions.validate_presence(value, **options)
+    end
+
+    it 'should define the method' do
+      expect(assertions)
+        .to respond_to(:validate_presence)
+        .with(1).argument
+        .and_keywords(:as, :message)
+    end
+
+    describe 'with nil' do
+      let(:value) { nil }
+
+      include_examples 'should raise an exception'
+    end
+
+    describe 'with an Object' do
+      let(:value) { Object.new.freeze }
+
+      include_examples 'should not raise an exception'
+    end
+
+    describe 'with an empty value' do
+      let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+      include_examples 'should raise an exception'
+    end
+
+    describe 'with a non-empty value' do
+      let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+      include_examples 'should not raise an exception'
+    end
+
+    wrap_context 'with as: value' do
+      let(:error_message) { "#{as} can't be blank" }
+
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
+
+        include_examples 'should not raise an exception'
+      end
+    end
+
+    wrap_context 'with message: value' do
+      describe 'with nil' do
+        let(:value) { nil }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with an Object' do
+        let(:value) { Object.new.freeze }
+
+        include_examples 'should not raise an exception'
+      end
+
+      describe 'with an empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(true) }
+
+        include_examples 'should raise an exception'
+      end
+
+      describe 'with a non-empty value' do
+        let(:value) { Struct.new(:empty) { def empty? = empty }.new(false) }
 
         include_examples 'should not raise an exception'
       end